+++ packages/components/src/modal/ModalDialog/ModalDialog.spec.js
@@ -73,6 +73,22 @@ describe('ModalDialog', () => {
 
       expect(component).toMatchSnapshot();
     });
+
+    it('with header divider if title is provided', () => {
+      component.setProps({ title: 'Some title' });
+      const header = component.find('.modal-header');
+
+      expect(header.hasClass('no-divider')).toBe(false);
+      expect(component).toMatchSnapshot();
+    });
+
+    it('without header divider if title is not provided', () => {
+      component.setProps({ title: null });
+      const header = component.find('.modal-header');
+
+      expect(header.hasClass('no-divider')).toBe(true);
+      expect(component).toMatchSnapshot();
+    });
   });
 
   describe('behaviourally', () => {